React to feedback
- pipe() called via Task.Run() prevents custom task scheduler from propagating
outside

- Replaced LimitedConcurrencyLevelTaskSchedule with built-in
ConcurrentExclusiveSchedulerPair

This lets us set a maximum concurrency via the ConcurrentScheduler, and all
channel interactions are scheduled via the ExclusiveScheduler, which ensures
only a single thread will be using it once.

Also, the way the prefetchCount is being set has been changed. In the
scenario of a single Consumer, this needs to be set to the passed-in
MaxConcurrency setting, or you don't get any concurrency at all.